update mapping logic, reduce state query
