+++ src/mappings/erc1155/handleURI.ts
@@ -1,29 +1,32 @@
 import { URILog } from '../../types/abi-interfaces/Erc1155';
 import { Erc1155__factory } from '../../types/contracts';
-import { Nft } from '../../types';
+import { Collection, Nft } from '../../types';
 import { getCollectionId, getNftId } from '../../utils/common';
 import { handleNetwork } from '../../utils/utilHandlers';
 import assert from 'assert';
 
 export async function handleERC1155Uri(event: URILog): Promise<void> {
-  let isERC1155 = false;
   const instance = Erc1155__factory.connect(event.address, api);
 
   const network = await handleNetwork(chainId);
+  const collectionId = getCollectionId(network.id, event.address);
 
-  try {
-    // https://eips.ethereum.org/EIPS/eip-1155#abstract
-    isERC1155 = await instance.supportsInterface('0xd9b67a26');
+  const collection = await Collection.get(collectionId);
 
-    if (!isERC1155) {
+  if (!collection) {
+    try {
+      // https://eips.ethereum.org/EIPS/eip-1155#abstract
+      const isERC1155 = await instance.supportsInterface('0xd9b67a26');
+
+      if (!isERC1155) {
+        return;
+      }
+    } catch (e) {
       return;
     }
-  } catch (e) {
-    return;
   }
-  assert(event.args, 'No event args on erc1155');
 
-  const collectionId = getCollectionId(network.id, event.address);
+  assert(event.args, 'No event args on erc1155');
   const nftId = getNftId(collectionId, event && event.args.id.toString());
 
   const nft = await Nft.get(nftId);
